Getting Supplier Codes only for active suppliers
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C28605

+++ TOneV2/Code/TOneV2/TOne.WhS.Routing.Data.SQL/SupplierZoneDetailsDataManager.cs
@@ -55,7 +55,7 @@ namespace TOne.WhS.Routing.Data.SQL
             return new SupplierZoneDetail()
             {
                 SupplierId = (int)reader["SupplierId"],
-                SupplierZoneId = (int)reader["SupplierZoneId"],
+                SupplierZoneId = (long)reader["SupplierZoneId"],
                 EffectiveRateValue = GetReaderValue<decimal>(reader, "EffectiveRateValue")
             };
         }
